auth/requirements.txt
asn1crypto==0.22.0
cffi==1.11.0
cryptography==2.0.3
Django==1.11.2
idna==2.6
pkg-resources==0.0.0
pycparser==2.18
PyJWT==1.5.3
pytz==2017.2
six==1.11.0
